created Technology component (work in progress)
